refactor: address additional PR review comments
- Filter MongoDB services at API level using serviceType parameter
- Move types to RealTimeSelectionForm.types.ts
- Improve comment about onSuccess: undefined
- Use Box component='li' instead of raw li element
- Use chipClasses and design tokens in ServiceOptionTag
