+++ ui/apps/pmm/src/pages/rta/selection/RealTimeSelectionForm.types.ts
@@ -1,0 +1,9 @@
+export interface ServiceOption {
+  type: 'cluster' | 'service';
+  id: string;
+  label: string;
+  serviceId?: string;
+  cluster?: string;
+}
+
+export type ClusterSelectionState = 'all' | 'partial' | 'none';
